feat: add edit and delete permissions in view

+++ blog/post/views.py
@@ -2,8 +2,8 @@ from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from .forms import PostCreateForm, CommentForm
-
 from .models import Post, Comment
+from django.core.exceptions import PermissionDenied
 
 def post_list(request):
     posts = Post.objects.all().order_by('-created_at')
@@ -67,6 +67,11 @@ def create_post(request):
 @login_required
 def edit_post(request, post_id):
     post = get_object_or_404(Post, pk=post_id, author=request.user)
+    
+    if not post.can_edit(request.user):
+        messages.error(request, 'нет прав для редактирования')
+        return redirect('post:post_detail', post_id=post.id)
+    
     if request.method == 'POST':
         # instance отвечает за передачу значений свойств объекта
         form = PostCreateForm(request.POST, instance=post)
@@ -88,6 +93,10 @@ def edit_post(request, post_id):
 @login_required
 def delete_post(request, post_id):
     post = get_object_or_404(Post, pk=post_id, author=request.user)
+    
+    if not post.can_delete(request.user):
+        raise PermissionDenied('У вас нет прав для удаления поста')
+        return redirect('post:post_detail', post_id=post.id)
     if request.method == 'POST':
         if 'confirm_delete' in request.POST:
             post.delete()
@@ -105,3 +114,24 @@ def delete_post(request, post_id):
         'page_title': f'Удаление {post.title}',
     }
     return render(request, 'post/post_details.html', context)
+
+
+@login_required
+def delete_comment(request, comment_id):
+    comment = get_object_or_404(Comment, pk=comment_id)
+    if not comment.can_delete(request.user):
+        raise PermissionDenied('Нет прав на удаление')
+
+    post_id = comment.post.id
+
+    if request.method == 'POST':
+        if 'confirm_delete_comment' in request.POST:
+            comment.delete()
+            messages.success('Комментарий успешно удален')
+        return redirect('post:post_detail', post_id=post_id)
+    
+    context = {
+        'comment': comment,
+        'page_title': 'Удаление комментария'
+    }
+    return render(request, 'post/comment_delete.html', context)
